lwio: error code from snprintf has a negative value. CID: 11706.
(lwio: r51481)

+++ lwio/server/srvstat/statistics.c
@@ -922,7 +922,7 @@ LwioSrvStatLogToString(
     {
         PSTR   pszCursor  = *ppszBuffer + *pulBytesUsed;
         size_t nBytesAvbl = *pulTotalLength - *pulBytesUsed;
-        size_t nWritten   = 0;
+        LONG   nWritten   = 0;
 
         switch (pValue->valueType)
         {
